+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unreleased
+### Fixed
+* Fixed an issue where some camdo models would not correctly remap material texture assignments.
+
 ## 0.11.1 - 2024-07-29
 ### Changed
 * Improved the error message when attempting to install the addon on an unsupported Blender version.