+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## unreleased
+## 0.13.0 - 2024-08-16
 ### Added
 * Added an option to replaces textures with assigned images from current scene for wimdo export.
 * Added an option to specify an image folder for replacing textures for wimdo export.